fix tests and move api
