+++ api/src/main/scala/hmda/api/http/InstitutionsHttpApi.scala
@@ -1,11 +1,11 @@
-package hmda.api.http.institutions
+package hmda.api.http
 
 import akka.actor.ActorSystem
 import akka.event.LoggingAdapter
 import akka.http.scaladsl.server.Directives._
 import akka.stream.ActorMaterializer
 import akka.util.Timeout
-import hmda.api.http.HmdaCustomDirectives
+import hmda.api.http.institutions.{ FilingPaths, InstitutionPaths, SubmissionPaths, UploadPaths }
 import hmda.api.protocol.processing.{ ApiErrorProtocol, InstitutionProtocol }
 
 trait InstitutionsHttpApi
