+++ persistence/src/test/scala/hmda/persistence/processing/LocalHmdaEventProcessorSpec.scala
@@ -98,6 +98,7 @@ class LocalHmdaEventProcessorSpec extends ActorSpec {
     }
 
     "process validation completed from event stream" in {
+      Thread.sleep(300)
       val msg = s"validation completed for submission $submissionId"
       checkEventStreamMessage(msg, ValidationCompleted(submissionId))
       checkSubmissionStatus(Validated)
